+++ Modules/Ruleset.Initialize/Public/Initialize-Module.ps1
@@ -138,7 +138,7 @@ function Initialize-Module
 			return $false
 		}
 
-		Write-Warning -Message $Message
+		Write-Warning -Message "[$($MyInvocation.InvocationName)] $Message"
 		return $true
 	}
 
@@ -200,7 +200,7 @@ function Initialize-Module
 					return $false
 				}
 
-				Write-Warning -Message $Message
+				Write-Warning -Message "[$($MyInvocation.InvocationName)] $Message"
 				return $true
 			}
 		}
@@ -231,7 +231,7 @@ function Initialize-Module
 				return $false
 			}
 
-			Write-Warning -Message $Message
+			Write-Warning -Message "[$($MyInvocation.InvocationName)] $Message"
 			return $true
 		}
 
@@ -352,7 +352,7 @@ function Initialize-Module
 				return $false
 			}
 
-			Write-Warning -Message $Message
+			Write-Warning -Message "[$($MyInvocation.InvocationName)] $Message"
 			return $true
 		}
 	}
@@ -418,7 +418,7 @@ function Initialize-Module
 			return $false
 		}
 
-		Write-Warning -Message $Message
+		Write-Warning -Message "[$($MyInvocation.InvocationName)] $Message"
 		return $true
 	}
 
@@ -571,7 +571,7 @@ function Initialize-Module
 		return $false
 	}
 
-	Write-Warning -Message $Message
+	Write-Warning -Message "[$($MyInvocation.InvocationName)] $Message"
 	return $true
 }