+++ Modules/Ruleset.Utility/Public/Set-Permission.ps1
@@ -321,7 +321,7 @@ function Set-Permission
 			{
 				try
 				{
-					Write-Warning -Message $_.Exception.Message
+					Write-Warning -Message "[$($MyInvocation.InvocationName)] $($_.Exception.Message)"
 					Set-Privilege SeSecurityPrivilege -ErrorAction Stop | Out-Null
 					Set-Acl -AclObject $Acl -LiteralPath $LiteralPath -ErrorAction Stop
 				}
@@ -346,7 +346,7 @@ function Set-Permission
 		{
 			try
 			{
-				Write-Warning -Message $_.Exception.Message
+				Write-Warning -Message "[$($MyInvocation.InvocationName)] $($_.Exception.Message)"
 				Set-Privilege SeSecurityPrivilege -ErrorAction Stop | Out-Null
 				Set-Acl -AclObject $Acl -LiteralPath $LiteralPath -ErrorAction Stop
 			}
@@ -397,7 +397,7 @@ function Set-Permission
 		{
 			try
 			{
-				Write-Warning -Message $_.Exception.Message
+				Write-Warning -Message "[$($MyInvocation.InvocationName)] $($_.Exception.Message)"
 				Set-Privilege SeSecurityPrivilege, SeTakeOwnershipPrivilege -ErrorAction Stop | Out-Null
 				Set-Acl -AclObject $Acl -LiteralPath $LiteralPath -ErrorAction Stop
 			}
@@ -469,7 +469,7 @@ function Set-Permission
 		{
 			try
 			{
-				Write-Warning -Message $_.Exception.Message
+				Write-Warning -Message "[$($MyInvocation.InvocationName)] $($_.Exception.Message)"
 				Set-Privilege SeSecurityPrivilege -ErrorAction Stop | Out-Null
 				Set-Acl -AclObject $Acl -LiteralPath $LiteralPath -ErrorAction Stop
 			}